Refactor to externalise functionality
